# backend/docs/swagger.yaml
basePath: /api/v1
...
info:
  contact: {}
  description: API do sistema de ressarcimento (SURE)
  title: SURE API
  version: "1.0"
paths:
...
  /api/v1/admin/prazos:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Lista configuraÃ§ão de prazos
      tags:
      - Admin
    post:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Salva configuraÃ§ão de prazos
      tags:
      - Admin
...
  /api/v1/events:
    get:
      produces:
      - text/event-stream
      responses:
        "200":
          description: stream
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: SSE global
      tags:
      - SSE
...
  /api/v1/processos/{id}/events:
    get:
      parameters:
      - description: ID do processo
        in: path
        name: id
        required: true
        type: integer
      produces:
      - text/event-stream
      responses:
        "200":
          description: stream
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: SSE de eventos por processo
      tags:
      - SSE
...
  /api/v1/processos/prazos:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Processos com prazo
      tags:
      - Processos
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"